+++ R/google_map_layer_geojson.R
@@ -140,7 +140,6 @@
 add_geojson <- function(map, data = get_map_data(map), layer_id = NULL, style = NULL, update_map_view = TRUE){
 
   ## TODO:
-  ## - drag & drop geojson - https://developers.google.com/maps/documentation/javascript/examples/layer-data-dragndrop
   ## - replicate blog: https://maps-apis.googleblog.com/2014/04/build-map-infographic-with-google-maps.html
 
   ## DataLayer events https://developers.google.com/maps/documentation/javascript/datalayer#add_event_handlers